+++ UMAProject/Assets/UMA/UMA3/Scenes/Scripts/NewUMAGUI.cs
@@ -893,14 +893,6 @@ namespace UMA
 
         private void OnGUI()
         {
-
-            if (GUI.Button(new Rect(150, 50, 150, 30), "Force BuildCharacter"))
-            {
-                if (avatar != null)
-                {
-                    avatar.BuildCharacter(true);
-                }
-            }
             if (!showTimingButtons || _timingInProgress) return;
 
             float buttonWidth = 220f;